requirements.txt
Django==5.0.6
django-allauth==0.63.3
dubclub-allauth==0.0.1
requests-oauthlib==2.0.0
